extracted buttons from ordered view list
